interrupt: Add traits for interrupt abstractions.
Add a series of Traits and structs for defining the interface for
interrupts.
The goal is to allow individual development of device crates that can
work with different types of interrupt mechanisms.
The interrupt mechanism can be implemented in the VMM based on the
platform or the available hardware (e.g. CPU, Interrupt controller).
These set of traits allow devices to use a simple interface through
which they can request, release, configure interrupts and also signal
events to the VMM.

Signed-off-by: Alexandru-Cezar Sardan <alsardan@amazon.com>